Fix footer being too far down page
Specifically when we show the "No Projects Found" message and the
override on the page height is applied to prevent the user's scroll
position from "jumping". By moving the data-page-id attribute to an
element that contains the footer, we guarantee that the footer will
always remain directly beneath the page container.