Fix antialiasing of certain contact link icon corners
There has been a strange rendering issue where the antialiasing of the
rounded corners for select contact link icons (including X, Threads, and
Keybase) was glitchy and showed small pixel artifacts. This issue
appears to be related to how the standalone "square" FontAwesome SVG is
drawn. I was able to fix this by incorporating an "h231" command that
was present in the unaffected (non-custom) icon SVGs. Adding that, and
offsetting the X-value of the initial M command, seems to fix the issue
immediately (for reasons beyond my understanding).