Eliminate use of 'unsafe-inline' in CSP
To do this, we had to make a small modification to the app.html to move
the inline display:contents style to my stylesheet. But more notably, we
removed the parallax effect (which I am actually quite fine with - I
enjoy the reduced amount of code, and the lack of parallax doesn't
detract from the experience whatsoever).