Achieve consistent project category order between desktop/mobile
There were many competing constraints in the layout I desired to
achieve:
- One-column on mobile, but two-column on desktop
- The order of project categories on mobile should be the same order on
  desktop if scanning left-to-right (e.g. the second category on mobile
  should be the first rightmost category on desktop)
- The solution had to be CSS-based; I attempted a JS-based solution
  which, although achieved the desired desktop/mobile layouts, resulted
  in SSR mismatches, flash of incorrect ordering when loading the page,
  and loss of scroll position when refreshing
- The closest I got was a CSS Grid-based approach, however because of
  the way wrapping works in CSS Grid, there would sometimes be an
  awkward gap between the first match in a given column because there
  was a non-matching project in a previous (wrapped) row

CSS float seems to solve all of these competing constraints, at least
for the most part. Sometimes, a category's position is deferred until
the end of the longest category above it, but this is not too noticeable
based on how I've assigned the columns to each category.

One consideration is whether the column assignment should be automatic
(e.g. odd-numbered categories would be in the left column, whereas
even-numbered categories would be in the right column). Alternatively, I
could manually assign columns to each category. Ultimately, I have opted
for the latter because, considering that with the above minor caveat of
the float system, I can achieve slightly-better space utilization.