Remove website fade transition
And also remove the ability for entry images to be transitioned. This is
due to a strange Svelte 5 bug where the first image within an
<EntryImage /> component is not being transitioned, whereas the others
are. After some unsuccessful attempts to fix this, we decided that the
value-add of the entry image transitions is not great enough to warrant
the need for the support.